test(scrape): tests Vitest des helpers + extraction module extractors.ts
Sépare les utilitaires purs de /api/scrape (SSRF guard, lookup CP/Ville,
sanitizers numériques, validation URL photo, détection site source) dans
app/api/scrape/extractors.ts pour les rendre testables sans monter Next/
Supabase/cheerio. route.ts importe désormais ces helpers depuis le module.

Ajoute tests/unit/scrape-extractors.test.ts (38 tests, 1 fichier) couvrant :
- isUrlAllowed : whitelist domaines, anti-SSRF (localhost, RFC1918, link-
  local, metadata cloud), anti-spoofing (pap.fr.evil.com), protocoles non
  HTTP, URLs malformées.
- detectSourceSite : leboncoin/seloger/pap/orpi/century21/laforet/figaro/
  bienici/logic-immo + fallback generic.
- findCityFromCP / findCPFromCity : Paris (1er-20e + 75116), Lyon, Marseille,
  DROM (Martinique, Guadeloupe, Réunion, Guyane, Mayotte), insensibilité
  à la casse et aux diacritiques (Saint-André ⇄ Saint-Andre).
- normalizeText : suppression diacritiques + lowercase + trim.
- cleanPrice/Surface/Rooms/DPE : plages, valeurs hors limites, falsy.
- isValidPhotoUrl : filtres logos/icons/pixels/SVG/GIF, exclusions LBC.
- Sanity checks sur les constantes (ALLOWED_DOMAINS, CP_TO_CITY DROM).

Note : un test documente la limitation connue de cleanDPE (matche la
PREMIÈRE lettre A-G venue, donc "Classe G" → "C"). Comportement laissé
tel quel : les callers en amont (JSON-LD, attrs data-*) passent déjà
des valeurs isolées. Refactor de la regex sortirait du périmètre du Lot 2.

Suite au rapport "Importer une annonce" : couvre la priorité #3
(coverage de tests sur les extracteurs).
